requirements.txt
aiogram>=3.22.0
alembic>=1.16.5
psycopg>=3.1.8
psycopg-binary>=3.1.8
pydantic-settings>=2.10.1
python-dateutil>=2.9.0.post0
requests>=2.32.5
SQLAlchemy>=2.0.43